Move UFS caching into underfs package
