+++ core/server/worker/src/main/java/alluxio/worker/page/PagedBlockStore.java
@@ -30,6 +30,7 @@ import alluxio.grpc.ErrorType;
 import alluxio.grpc.UfsReadOptions;
 import alluxio.proto.dataserver.Protocol;
 import alluxio.resource.LockResource;
+import alluxio.underfs.UfsInputStreamCache;
 import alluxio.underfs.UfsManager;
 import alluxio.worker.block.AllocateOptions;
 import alluxio.worker.block.BlockLock;
@@ -42,7 +43,6 @@ import alluxio.worker.block.BlockStoreEventListener;
 import alluxio.worker.block.BlockStoreLocation;
 import alluxio.worker.block.BlockStoreMeta;
 import alluxio.worker.block.CreateBlockOptions;
-import alluxio.worker.block.UfsInputStreamCache;
 import alluxio.worker.block.io.BlockReader;
 import alluxio.worker.block.io.BlockWriter;
 import alluxio.worker.block.io.DelegatingBlockReader;
